reportevent权限修改，d.ts文件补全，js用例补全

+++ services/common/src/bundle_active_core.cpp
@@ -407,7 +407,7 @@ void BundleActiveCore::OnUserSwitched(const int userId)
 
 int BundleActiveCore::ReportEvent(BundleActiveEvent& event, const int userId)
 {
-    BUNDLE_ACTIVE_LOGI("FLUSH interval is %{public}lld, debug is %{public}d", flushInterval_, debugCore_);
+    BUNDLE_ACTIVE_LOGI("FLUSH interval is %{public}lld, debug is %{public}d", (long long)flushInterval_, debugCore_);
     if (debugCore_) {
         event.PrintEvent();
     }
